+++ packages/api/cms-api/src/user-permissions/user-permissions.types.ts
@@ -84,3 +84,16 @@ export interface UserPermissionsModuleAsyncOptions extends Pick<ModuleMetadata, 
 export const CombinedPermission: Record<string, string> = {
     ...CorePermission,
 };
+
+/**
+ * Helper to register additional permissions into the permission enum used for the GraphQL schema.
+ * Only use this if you're building a library that requires additional permissions.
+ * For application-level permissions, use the `AppPermission` option in the module registration methods.
+ *
+ * @param additionalPermissions
+ */
+export const registerAdditionalPermissions = (additionalPermissions: Record<string, string>): void => {
+    Object.entries(additionalPermissions).forEach(([key, value]) => {
+        CombinedPermission[key] = value;
+    });
+};