+++ packages/mail-react/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @comet/mail-react
 
+## 8.11.1
+
 ## 8.11.0
 
 ### Minor Changes